Bug Fix: Bowling too slow no longer ends frame as soon as rou reach the pins
There was an on trigger enter on the pinsetter that flagged the ball as out of play when it entered. This was to speed up the time it takes to calculate pinfall. I changed it to an ontriggerexit wich prevents the game from ending premeturly but does make the calculation take slightly longer in normal use.