Attempted to add spin
Even incredible torque makes almost no difference if the drag is set to low, and if I increase the drag the ball slows down to much, cannot add spin to ball using physics engine. I dont think I want to add spin if it involves path following for this game.